main.py
#!/usr/bin/env python3
"""
RSS Slack Consolidator - Main Entry Point

Consolidates RSS feeds and posts relevant items to Slack.
"""

import logging
import os
import sys
from pathlib import Path

from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Setup logging
log_dir = Path("logs")
log_dir.mkdir(exist_ok=True)

logging.basicConfig(
    level=getattr(logging, os.getenv("LOG_LEVEL", "INFO")),
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('logs/app.log'),
        logging.StreamHandler(sys.stdout)
    ]
)

logger = logging.getLogger(__name__)


def check_environment():
    """Verify required environment variables are set."""
    required_vars = ["SLACK_BOT_TOKEN", "SLACK_CHANNEL_ID"]
    missing = [var for var in required_vars if not os.getenv(var)]

    if missing:
        logger.error(f"Missing required environment variables: {', '.join(missing)}")
        logger.error("Please copy .env.example to .env and configure your settings")
        return False

    return True


def main():
    """Main application entry point."""
    logger.info("Starting RSS Slack Consolidator...")

    # Check environment configuration
    if not check_environment():
        sys.exit(1)

    # TODO: Initialize database
    # from src.database import init_db
    # init_db()

    # TODO: Set up scheduler
    # from src.scheduler import start_scheduler
    # start_scheduler()

    logger.info("RSS Slack Consolidator is running")
    logger.info("Press Ctrl+C to stop")

    # TODO: Keep the application running
    # try:
    #     while True:
    #         time.sleep(1)
    # except KeyboardInterrupt:
    #     logger.info("Shutting down gracefully...")

    logger.warning("Implementation in progress - see CLAUDE.md for development guide")
    logger.info("Next steps:")
    logger.info("1. Implement src/database.py (use database-specialist)")
    logger.info("2. Implement src/feed_parser.py (use general-assistant)")
    logger.info("3. Implement src/slack_poster.py (use general-assistant)")
    logger.info("4. Implement src/scheduler.py (use general-assistant)")
    logger.info("5. Implement src/filters.py (use general-assistant)")


if __name__ == "__main__":
    main()
